+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add MultiLogger to support logging to multiple Logger instances
 - Add resize listener on HTMLVideoElement in demo
 - Add simulcast integration tests
+- Add the device change event controller
 
 ### Changed
 - Use GET instead of POST to obtain TURN credentials